[save-images] Make all threads exception-safe
Ensure errors are re-raised safely from worker threads by using non-blocking
puts and monitoring a common error queue.
